[core] Move mbgl::Pass and mbgl::PassRefPtr to a separate header
